+++ src/app/(blog)/layout.tsx
@@ -1,7 +1,7 @@
 import { Geist, Geist_Mono } from "next/font/google";
 import "./globals.css";
-import Header from "./Header";
-import Footer from "./Footer";
+import Header from "../Header";
+import Footer from "../Footer";
 
 const geistSans = Geist({
   variable: "--font-geist-sans",
@@ -23,11 +23,25 @@ export default function RootLayout({
       <body
         className={`${geistSans.variable} ${geistMono.variable} antialiased flex flex-col min-h-screen bg-white`}
       >
-        <Header />
+        <Header
+          fromColor="slate-300"
+          toColor="slate-500"
+          logo1="/LOGO1.gif"
+          logo1Link="/"
+          logo2="/LOGO_BLOG.png"
+          logo2Link="/blog"
+          textColor="black"
+          textHoverColor="gray-700"
+        />
         <main className="flex-grow">
           <div className="max-w-[1280px] mx-auto pt-[80px]">{children}</div>
         </main>
-        <Footer />
+        <Footer
+          fromColor="slate-300"
+          toColor="slate-500"
+          textColor="black"
+          textHoverColor="gray-700"
+        />
       </body>
     </html>
   );